Replace TSC code with clock_gettime(CLOCK_MONOTONIC) on Linux
On the raw C-level TSC is faster than clock_gettime() for each
individual call. However we need to calibrate the process by running a
5ms benchmark per core. The small performance boost is lost this way.
On top of that the PHP overhead in each function makes the small
different in C-level unimportant in comparison.

Removing TSC allows us to remove all the nasty invariant, frequency and
affinity code, leaving us with a much simpler and more precise solution.